+++ src/components/WeekDayCalendar.tsx
@@ -1,17 +1,40 @@
 import moment, { months } from 'moment';
 import React, { Dispatch, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { Action } from 'redux';
-import { addAssignedHour, initialStateProp, setCurrentWeek, setnewCurrentWeekMoment } from '../redux/reducers/dayReducer';
-import { RootState } from '../redux/store';
-import { AiFillCaretLeft, AiFillCaretRight } from "react-icons/ai";
-import { addNewMonth, setInitalActiveWeekInd } from '../redux/reducers/monthReducer';
+// import { Action } from 'redux';
+// import { addAssignedHour, initialStateProp, setCurrentWeek, setnewCurrentWeekMoment } from '../redux/reducers/dayReducer';
+// import { RootState } from '../redux/store';
+// import { AiFillCaretLeft, AiFillCaretRight } from "react-icons/ai";
+// import { addNewMonth, setInitalActiveWeekInd } from '../redux/reducers/monthReducer';
 
 const WeekDayCalendar: React.FC = () => {
   const dispatch = useDispatch()
-  const { activeWeek, hoursOfDay} = useSelector((state: RootState) => state.week_days)
-  const { years, currentMonth, initialActiveWeek, activeMonth, activeYear } = useSelector((state: RootState) => state.month)
-  
+  // const { activeWeek, hoursOfDay} = useSelector((state: RootState) => state.week_days)
+  moment.updateLocale('en', { week: { dow: 1 } })
+  // const { years, currentMonth, initialActiveWeek, activeMonth, activeYear } = useSelector((state: RootState) => state.month)
+  let currentWeekMoment = moment().startOf('week')
+  const genWeek = (currentWeekMoment) => {
+    let endDay = ''
+    let binDay = `${currentWeekMoment.format('DD')}-${currentWeekMoment.format('MMM')}`
+     let initalDays = []
+    
+     for (let j = 0; j < 7; j++) {
+            let day = {
+                day: currentWeekMoment.format('dddd'),
+                date: currentWeekMoment.clone(),
+                assignedHours: []
+            }
+       let index = `${currentWeekMoment.format('DD')}-${currentWeekMoment.format('MMM')}`
+            initalDays.push(day);
+       currentWeekMoment.add(1, 'day');
+     }
+    return {[binDay]: initalDays}
+  }
+
+  const days = genWeek(currentWeekMoment)
+  const days1 = genWeek(currentWeekMoment)
+  console.log(days, 'days');
+  console.log(days1, 'days1');
 
  const handleDrop = (e) => {
     // let user = JSON.parse(e.dataTransfer.getData("user"));
@@ -37,13 +60,13 @@ const WeekDayCalendar: React.FC = () => {
 
   }
 
-  const currentYears = years[activeYear][activeMonth][initialActiveWeek] 
+//   const currentYears = years[activeYear][activeMonth][initialActiveWeek] 
 
-console.log(currentYears, initialActiveWeek);
+// console.log(years, initialActiveWeek);
 
   return (
     <table className='border'>
-      <thead className='border-b'>
+      {/* <thead className='border-b'>
         <tr className='border font-bold text-center px-2'>{currentYears.currentWeekMoment?.format('MMMM')} - {currentYears.currentWeekMoment?.format("YY")}</tr>
         <tr className='border px-2'>
           <th><button onClick={() => {
@@ -79,8 +102,8 @@ console.log(currentYears, initialActiveWeek);
               </td>)}
           </tr>
         ))}
-      </tbody>
-
+      </tbody> */}
+some table
     </table>
   );
 };
