+++ src/redux/reducers/monthReducer.ts
@@ -2,106 +2,55 @@ import { createSlice } from '@reduxjs/toolkit';
 import moment from 'moment';
 import { weekInterface } from './dayReducer';
 
-
 export interface yearInterface {
-    [year: string]: { [month: string]: weekInterface[] }
+    [year: string]: weekInterface[]
 }
 
-
-export interface initialMonthProp {
+export interface yearStateInterface {
     years: yearInterface
     currentMonth: weekInterface[],
-    activeYear: string,
-    activeMonth: string,
-    initialActiveWeek: number
 }
+
+
+const genWeek = (currentWeekMoment: moment.Moment) => {
+    let binDay = `${currentWeekMoment.format('DD')}-${currentWeekMoment.format('MMM')}`
+    let initalDays = []
+
+    for (let j = 0; j < 7; j++) {
+        let day = {
+            day: currentWeekMoment.format('dddd'),
+            date: currentWeekMoment.clone(),
+            assignedHours: []
+        }
+        initalDays.push(day);
+        currentWeekMoment.add(1, 'day');
+    }
+    return { binDay, initalDays }
+}
+
+const genMonth = (monthMoment: moment.Moment) => {
+    let i = 0
+    let monthArry = []
+    while (i < 6) {
+        const week = genWeek(monthMoment.startOf('week'))
+        monthArry.push(week)
+    }
+    return monthArry
+}
+
+const motMoment = moment().startOf('month')
+const initalYear = motMoment.format("yyyy")
+const monthArray = genMonth(motMoment)
+let yearsCont = {}
+monthArray.map(mth => {
+    yearsCont[mth.binDay] = mth.initalDays
+})
+const initialState: yearStateInterface = {
+    years: {
+        [initalYear]: yearsCont
+    }
+}
+
 
 moment.updateLocale('en', { week: { dow: 1 } })
-// Get the current month and calculate the number of weeks in it
-let currentMonth = moment().startOf('month');
-// let weeksInMonth = currentMonth.isoWeeksInMonth();
 
-
-export const generateMonth = (currentMonth: moment.Moment, initalWeekMoment: moment.Moment) => {
-    let initialActiveInd = 0
-    let monthWeeks = []
-    for (let i = 0; i < 6; i++) {
-        let currentWeekStart = currentMonth.clone().startOf('week');
-        let initalDays = []
-
-        for (let j = 0; j < 6; j++) {
-            let day = {
-                day: currentWeekStart.format('dddd'),
-                date: currentWeekStart.clone(),
-                assignedHours: []
-            }
-            initalDays.push(day);
-            currentWeekStart.add(1, 'day');
-        }
-
-        monthWeeks.push({
-            daysOfWeek: initalDays,
-            currentWeekMoment: currentWeekStart.clone()
-        });
-        if (initalWeekMoment.isSame(currentWeekStart, 'week')) {
-            initialActiveInd = i
-        }
-        currentMonth.add(1, 'week');
-    }
-    return { monthWeeks, initialActiveInd }
-}
-
-const { monthWeeks, initialActiveInd } = generateMonth(moment().startOf('month'), moment().startOf('week'))
-
-let initialYear = moment().format('yyyy').toString()
-let initialMonth = moment().format('MMMM').toString()
-
-const initialState: initialMonthProp = {
-    years: {
-        [initialYear]: {
-            [initialMonth]: monthWeeks,
-        },
-    },
-    currentMonth: monthWeeks,
-    activeYear: moment().format("YYYY"),
-    activeMonth: moment().format("MMMM").toString(),
-    initialActiveWeek: initialActiveInd
-}
-
-
-export const monthSlice = createSlice({
-    name: 'month',
-    initialState,
-    reducers: {
-        addNewMonth: (state, action) => {
-            let curmonth = action.payload.month.clone().format("MMMM")
-            let curYear = action.payload.week.clone().format("yy")
-            console.log(state.years[curYear]);
-
-            if (state.years[curYear] && state.years[curYear][curmonth]) {
-                state.activeMonth = curmonth
-                state.activeYear = curYear
-                return
-            }
-            const { monthWeeks, initialActiveInd } = generateMonth(action.payload.month, action.payload.week)
-            console.log(curYear, curmonth);
-
-            state.years[curYear] = {
-                [curmonth]: monthWeeks
-            }
-            state.initialActiveWeek = initialActiveInd
-            state.activeMonth = curmonth
-            state.activeYear = curYear
-        },
-        setCurrentMonth: (state, action) => {
-            state.currentMonth = action.payload
-        },
-        setInitalActiveWeekInd: (state, action) => {
-            state.initialActiveWeek = action.payload
-        }
-    },
-})
-
-export const { setCurrentMonth, setInitalActiveWeekInd, addNewMonth } = monthSlice.actions
-
-export default monthSlice.reducer
